Remove telemetry modules
