+++ docs/source/logging.md
@@ -53,9 +53,8 @@ correlated across stages. Tool executions follow the same pattern. When a tool
 starts, finishes or fails, a structured log entry is emitted describing the
 result.
 
-Metrics are recorded alongside these logs. Duration and error counts are stored
-in ``context.metrics`` and can be exported to Prometheus using
-``pipeline.observability.MetricsServerManager.start()``.
+Metrics collection is currently disabled while the telemetry architecture is being
+reworked.
 
 For production deployments, `config/logging_prod.yaml` contains a recommended
 configuration enabling JSON formatted logs and file rotation.
