+++ src/entity/core/plugins/base.py
@@ -51,9 +51,6 @@ class BasePlugin:
         start = time.perf_counter()
         response = await context.call_llm(context, prompt, purpose=purpose)
         duration = time.perf_counter() - start
-        if context.get_failure_info() is None and hasattr(context._state, "metrics"):
-            key = f"{context.current_stage}:{self.__class__.__name__}:{purpose}"
-            context._state.metrics.record_llm_duration(key, duration)
         self.logger.info(
             "LLM call completed",
             extra={
