--- src/pipeline/observability/__init__.py
@@ -1,9 +1,0 @@
-from .metrics import MetricsServer, MetricsServerManager
-from .tracing import start_span, traced
-
-__all__ = [
-    "MetricsServer",
-    "MetricsServerManager",
-    "start_span",
-    "traced",
-]
