+++ tests/test_llm_helpers.py
@@ -4,7 +4,6 @@ from datetime import datetime
 import pytest
 
 from entity.core.resources.container import ResourceContainer
-from entity.core.state import MetricsCollector
 from pipeline import (
     ConversationEntry,
     PipelineStage,
@@ -29,7 +28,6 @@ def make_context(llm=None) -> PluginContext:
             ConversationEntry(content="hi", role="user", timestamp=datetime.now())
         ],
         pipeline_id="1",
-        metrics=MetricsCollector(),
     )
     resources = ResourceContainer()
     if llm is not None:
