--- tests/test_observability.py
@@ -1,43 +1,0 @@
-import asyncio
-import time
-
-from pipeline import (
-    PipelineStage,
-    PluginRegistry,
-    PromptPlugin,
-    SystemRegistries,
-    ToolRegistry,
-    execute_pipeline,
-)
-
-from entity.core.resources.container import ResourceContainer
-
-
-class TimedPlugin(PromptPlugin):
-    stages = [PipelineStage.DELIVER]
-
-    async def _execute_impl(self, context):
-        await asyncio.sleep(0.05)
-        context.set_response("ok")
-
-
-def make_capabilities():
-    plugins = PluginRegistry()
-    asyncio.run(
-        plugins.register_plugin_for_stage(TimedPlugin({}), PipelineStage.DELIVER)
-    )
-    return SystemRegistries(ResourceContainer(), ToolRegistry(), plugins)
-
-
-def test_metrics_overhead():
-    capabilities = make_capabilities()
-    start = time.time()
-    result, metrics = asyncio.run(
-        execute_pipeline("hi", capabilities, return_metrics=True)
-    )
-    duration = time.time() - start
-    plugin_key = f"{PipelineStage.DELIVER}:TimedPlugin"
-    recorded = metrics.plugin_durations[plugin_key][0]
-    assert result == "ok"
-    assert recorded >= 0.05
-    assert duration - recorded < 0.05
