+++ tests/test_react_prompt.py
@@ -2,7 +2,6 @@ import asyncio
 from datetime import datetime
 
 from entity.core.resources.container import ResourceContainer
-from entity.core.state import MetricsCollector
 from pipeline import (
     ConversationEntry,
     PipelineState,
@@ -33,7 +32,6 @@ def make_context(llm: FakeLLM):
             )
         ],
         pipeline_id="1",
-        metrics=MetricsCollector(),
     )
     resources = ResourceContainer()
     tools = ToolRegistry()
