--- user_plugins/examples/observability_metrics.py
@@ -1,29 +1,0 @@
-"""Start Prometheus metrics and trace a simple asynchronous task.
-
-Run with ``python -m examples.observability_metrics`` or install the package in
-editable mode.
-"""
-
-from __future__ import annotations
-
-import asyncio
-
-from .utilities import enable_plugins_namespace
-
-enable_plugins_namespace()
-
-from pipeline.observability import MetricsServerManager, start_span
-
-
-async def sample_task() -> None:
-    async with start_span("sample_task"):
-        await asyncio.sleep(0.2)
-
-
-async def main() -> None:
-    MetricsServerManager.start(port=9001)
-    await sample_task()
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
